@@ -9,7 +9,7 @@ import Foundation
 ///
 /// Data is persisted to a file located in the app's documents folder using JSON encoding and decoding.
 /// All read and write operations are performed asynchronously on a dedicated background queue.
-/// 
+///
 public class CodablePersistence {
 
     private let queue = DispatchQueue(label: "codablepersistence.data", qos: .userInitiated)
@@ -29,7 +29,7 @@ public class CodablePersistence {
     /// - Parameters:
     ///   - value: The default value to return if loading or decoding fails.
     ///   - completion: A closure called on the main thread with the loaded or default value.
-    public func load<T: Decodable>(default value: T, completion: @escaping (T) -> Void) {
+    public func load<T: Decodable & Sendable>(default value: T, completion: @Sendable @escaping (T) -> Void) {
         queue.async { [url] in
             var data: T = value
             do {
@@ -49,7 +49,7 @@ public class CodablePersistence {
     /// Asynchronously saves codable data to a file in the documents directory.
     ///
     /// - Parameter data: The data to be encoded and saved.
-    public func save<T: Codable>(_ data: T) {
+    public func save<T: Codable & Sendable>(_ data: T) {
         queue.async { [url] in
             do {
                 let json = try JSONEncoder().encode(data)
